requirements.txt
# --- Core ---
reflex>=0.6.0
python-dotenv

# --- Backend & Data ---
supabase
google-cloud-secret-manager
requests

# --- Async Task Queue ---
celery[redis]>=5.3.0
redis>=5.0.0